[serverless] log writer under grpc reporter
